pyproject.toml
[tool.poetry]
name = "busca-vagas-serpapi"
version = "0.1.0"
description = ""
authors = ["eng-alecrim <lucasalecrimvieira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-search-results = "^2.4.2"
pydantic = "^2.8.2"
pandas = "^2.2.2"
pyarrow = "^17.0.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.1"
ignr = "^2.2"
black = "^24.4.2"
jupyterlab = "^4.2.3"
catppuccin-jupyterlab = "^0.2.2"

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]
ignore = ["E501", "E731", "PLR1711"]

[tool.ruff.format]
indent-style = "space"
preview = true
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.taskipy.tasks]
lint = "ruff check . && ruff check . --diff"
format = "ruff check . --fix && ruff format ."
pre_test = "task lint"
test = "pytest -s -x --cov=busca_vagas_serpapi -vv"
post_test = "coverage html"
jupyter = "python -m jupyterlab"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
